avancement du tracking

+++ src/testMath/testTraitement1.java
@@ -37,11 +37,11 @@ public class testTraitement1 {
 		VirtualPixel[][] testscreen = new VirtualPixel[LENGHT][HEIGHT];
 		for (int i = 0; i < LENGHT; i++) {
 			for (int j = 0; j < HEIGHT; j++) {
-				testscreen[i][j] = new VirtualPixel(false, 0, new Point(i, j),0);
+				testscreen[i][j] = new VirtualPixel(false, 0, new Point(i, j), (byte) 0);
 			}
 		}
 		if ((x <= LENGHT) && (y <= HEIGHT)) {
-			testscreen[x][y] = new VirtualPixel(true, 0, new Point(x, y),0);
+			testscreen[x][y] = new VirtualPixel(true, 0, new Point(x, y),(byte) 0);
 			traitement.updateConnexe(testscreen, LENGHT, HEIGHT);
 			Point pos = traitement.getGroupePos(1);
 			if ((x == (int) pos.getX()) && (y == (int) pos.getY())) {
